rename MobileMenu so there are not two compones with same name
